+++ collectivo/collectivo/composables/directus.ts
@@ -5,7 +5,7 @@ export const useDirectus = () => {
 
 export function requireAuth() {
   const directus = useDirectus();
-  const user = useUser();
+  const user = useCollectivoUser();
   const runtimeConfig = useRuntimeConfig();
 
   // If user is authenticated, do nothing
@@ -17,12 +17,11 @@ export function requireAuth() {
   if (runtimeConfig.public.authService === "keycloak") {
     return navigateTo(
       `${runtimeConfig.public.directusUrl}/auth/login/keycloak?redirect=${runtimeConfig.public.collectivoUrl}`,
-      { external: true },
+      { external: true }
     );
   } else {
     throw new Error(
-      "Unknown auth service in nuxt.config: " +
-        runtimeConfig.public.authService,
+      "Unknown auth service in nuxt.config: " + runtimeConfig.public.authService
     );
   }
 }
