+++ collectivo/collectivo/layouts/components/Sidebar.vue
@@ -4,7 +4,7 @@ import MenuItem from "./MenuItem.vue";
 import pkg from "../../package.json";
 
 const menus = useCollectivoMenus();
-const user = useUser();
+const user = useCollectivoUser();
 
 const mainMenuItems = Object.values(menus.value.main).sort(
   (a, b) => (a.order ?? 100) - (b.order ?? 100)
