+++ collectivo/collectivo/plugins/directus.client.ts
@@ -7,12 +7,12 @@ export default defineNuxtPlugin({
   async setup() {
     let directus;
     const runtimeConfig = useRuntimeConfig();
-    const user = useUser();
+    const user = useCollectivoUser();
 
     // Create directus REST client or redirect to offline error page
     try {
       directus = createDirectus<CollectivoSchema>(
-        runtimeConfig.public.directusUrl as string,
+        runtimeConfig.public.directusUrl as string
       )
         .with(authentication("cookie", { credentials: "include" }))
         .with(rest({ credentials: "include" }));
